@@ -14,8 +14,8 @@ repositories {
 	mavenCentral()
 }
 
-sourceCompatibility = JavaVersion.VERSION_14
-targetCompatibility = JavaVersion.VERSION_14
+sourceCompatibility = JavaVersion.VERSION_15
+targetCompatibility = JavaVersion.VERSION_15
 
 dependencies {
 	implementation project(':viperpit-commons')
