@@ -4,8 +4,8 @@ plugins {
 	id 'io.spring.dependency-management'
 }
 
-sourceCompatibility = JavaVersion.VERSION_14
-targetCompatibility = JavaVersion.VERSION_14
+sourceCompatibility = JavaVersion.VERSION_15
+targetCompatibility = JavaVersion.VERSION_15
 
 dependencies {
 	implementation project(':viperpit-commons')
